f3probe: fix allocation error of safe device
probe_device_max_blocks() was returning an int that once multiplied
by block_size was out range.
This patch just enforces probe_device_max_blocks() to return
an uint64_t, and does the needed adjustments in the code.

+++ libdevs.c
@@ -1209,7 +1209,7 @@ static const char *sdev_get_filename(struct device *dev)
 	return dev_get_filename(dev_sdev(dev)->shadow_dev);
 }
 
-struct device *create_safe_device(struct device *dev, int max_blocks,
+struct device *create_safe_device(struct device *dev, uint64_t max_blocks,
 	int min_memory)
 {
 	struct safe_device *sdev;
