+++ libprobe.h
@@ -5,7 +5,7 @@
 
 #include "libdevs.h"
 
-int probe_device_max_blocks(struct device *dev);
+uint64_t probe_device_max_blocks(struct device *dev);
 
 int probe_device(struct device *dev, uint64_t *preal_size_byte,
 	uint64_t *pannounced_size_byte, int *pwrap,
